+++ cli/src/ui/editors.rs
@@ -11,9 +11,16 @@ use super::{
     util::{default_style, highlight_style},
 };
 
+mod key_usage;
 pub use key_usage::KeyUsageEditor;
+
+mod node_id;
 pub use node_id::NodeIdEditor;
-pub use permission::PermissionEditor;
+
+mod permissions;
+pub use permissions::PermissionsEditor;
+
+mod validity_period;
 pub use validity_period::ValidityPeriodEditor;
 
 pub enum EditorEventResult {
@@ -185,664 +192,5 @@ fn adjust_render_area(area: &Rect, text: &String) -> Option<Rect> {
         Some(adjusted_area)
     } else {
         None
-    }
-}
-
-mod key_usage {
-    use super::*;
-
-    use std::collections::HashSet;
-
-    use golem_certificate::schemas::certificate::key_usage::{KeyUsage, Usage};
-
-    const KEY_USAGE_ORDER: [(Usage, &str); 3] = [
-        (Usage::SignCertificate, "Sign certificate"),
-        (Usage::SignManifest, "Sign manifest"),
-        (Usage::SignNode, "Sign node"),
-    ];
-
-    pub struct KeyUsageEditor {
-        highlight: Option<usize>,
-        key_usage: KeyUsage,
-        usage: HashSet<Usage>,
-    }
-
-    impl KeyUsageEditor {
-        pub fn new(key_usage: Option<KeyUsage>) -> Self {
-            let key_usage = key_usage.unwrap_or(KeyUsage::Limited([Usage::SignNode].into()));
-            let usage = match &key_usage {
-                KeyUsage::All => HashSet::new(),
-                KeyUsage::Limited(usage) => usage.iter().map(|usage| usage.to_owned()).collect(),
-            };
-            Self {
-                highlight: None,
-                key_usage,
-                usage,
-            }
-        }
-
-        pub fn get_key_usage(&self) -> KeyUsage {
-            match &self.key_usage {
-                KeyUsage::All => KeyUsage::All,
-                KeyUsage::Limited(_) => KeyUsage::Limited(self.usage.iter().map(|usage| usage.to_owned()).collect()),
-            }
-        }
-    }
-
-    impl Default for KeyUsageEditor {
-        fn default() -> Self {
-            Self::new(None)
-        }
-    }
-
-    impl EditorComponent for KeyUsageEditor {
-        fn enter_from_below(&mut self) {
-            self.highlight = Some(self.calculate_render_height() - 1);
-        }
-
-        fn enter_from_top(&mut self) {
-            self.highlight = Some(0);
-        }
-
-        fn get_highlight(&self) -> Option<usize> {
-            self.highlight
-        }
-
-        fn handle_key_event(&mut self, key_event: KeyEvent) -> EditorEventResult {
-            if let Some(highlight) = self.highlight {
-                match key_event.code {
-                    KeyCode::Esc => EditorEventResult::Escaped,
-                    KeyCode::Up =>
-                        if highlight > 0 {
-                            self.highlight = Some(highlight - 1);
-                            EditorEventResult::KeepActive
-                        } else {
-                            self.highlight = None;
-                            EditorEventResult::ExitTop
-                        }
-                    KeyCode::Down =>
-                        if highlight < self.calculate_render_height() - 1 {
-                            self.highlight = Some(highlight + 1);
-                            EditorEventResult::KeepActive
-                        } else {
-                            self.highlight = None;
-                            EditorEventResult::ExitBottom
-                        }
-                    KeyCode::Enter => {
-                            if highlight == 0 {
-                                self.key_usage = match &self.key_usage {
-                                    KeyUsage::All => KeyUsage::Limited(Default::default()),
-                                    KeyUsage::Limited(_) => KeyUsage::All,
-                                };
-                            } else if highlight > 0 && highlight < self.calculate_render_height() {
-                                let idx = highlight - 1;
-                                let usage = KEY_USAGE_ORDER[idx].0.clone();
-                                if self.usage.contains(&usage) {
-                                    self.usage.remove(&usage);
-                                } else {
-                                    self.usage.insert(usage);
-                                }
-                            }
-                            EditorEventResult::KeepActive
-                        }
-                    _ => EditorEventResult::KeepActive,
-                }
-            } else {
-                EditorEventResult::Inactive
-            }
-        }
-
-        fn calculate_render_height(&self) -> usize {
-            match &self.key_usage {
-                KeyUsage::All => 1,
-                KeyUsage::Limited(_) => 1 + KEY_USAGE_ORDER.len(),
-            }
-        }
-
-        fn get_text_output(&self, text: &mut String) {
-            write!(text, "Key usage:").unwrap();
-            match self.key_usage {
-                KeyUsage::All => write!(text, " All").unwrap(),
-                KeyUsage::Limited(_) => {
-                    writeln!(text, "").unwrap();
-                    KEY_USAGE_ORDER.iter().for_each(|(usage, usage_str)| {
-                        writeln!(text, "  [{}] {}", if self.usage.contains(usage) { "*" } else { " " }, usage_str).unwrap();
-                    });
-                },
-            }
-        }
-
-        fn get_highlight_prefix(&self) -> Option<usize> {
-            self.highlight.map(|highlight| if highlight == 0 { 0 } else { 2 })
-        }
-
-        fn get_editor(&mut self) -> Option<&mut TextInput> {
-            None
-        }
-
-        fn get_parse_error(&mut self) -> Option<&mut ModalMessage> {
-            None
-        }
-    }
-}
-
-mod node_id {
-    use super::*;
-
-    use std::str::FromStr;
-
-    use ya_client_model::NodeId;
-
-    pub struct NodeIdEditor {
-        node_id: String,
-        highlight: bool,
-        editor: Option<TextInput>,
-        parse_error: Option<ModalMessage>,
-    }
-
-    impl NodeIdEditor {
-        pub fn new(node_id: Option<NodeId>) -> Self {
-            Self {
-                node_id: node_id.map(|id| id.to_string()).unwrap_or(String::from("0x0000000000000000000000000000000000000000")),
-                highlight: false,
-                editor: None,
-                parse_error: None,
-            }
-        }
-
-        pub fn get_node_id(&self) -> NodeId {
-            NodeId::from_str(&self.node_id).unwrap()
-        }
-    }
-
-    impl Default for NodeIdEditor {
-        fn default() -> Self {
-            Self::new(None)
-        }
-    }
-
-    impl EditorComponent for NodeIdEditor {
-        fn enter_from_below(&mut self) {
-            self.highlight = true;
-        }
-
-        fn enter_from_top(&mut self) {
-            self.highlight = true;
-        }
-
-        fn get_highlight(&self) -> Option<usize> {
-            if self.highlight {
-                Some(0)
-            } else {
-                None
-            }
-        }
-
-        fn handle_key_event(&mut self, key_event: KeyEvent) -> EditorEventResult {
-            if let Some(parse_error) = self.parse_error.as_mut() {
-                match parse_error.handle_key_event(key_event) {
-                    Ok(status) => match status {
-                        ComponentStatus::Active => {},
-                        _ => self.parse_error = None,
-                    }
-                    Err(_) => {},
-                }
-                EditorEventResult::KeepActive
-            } else if let Some(editor) = self.editor.as_mut() {
-                match editor.handle_key_event(key_event) {
-                    Ok(status) => match status {
-                        ComponentStatus::Active => {},
-                        ComponentStatus::Closed => {
-                            let node_id_string = {
-                                let mut s = editor.get_text().to_owned();
-                                if s.len() == 40 {
-                                    s = format!("0x{}", s);
-                                }
-                                s
-                            };
-                            match NodeId::from_str(&node_id_string) {
-                                Ok(_) => {
-                                    self.node_id = node_id_string;
-                                    self.editor = None;
-                                }
-                                Err(err) => {
-                                    let parse_error = ModalMessage::new("NodeId parse error", err.to_string());
-                                    self.parse_error = Some(parse_error);
-                                },
-                            }
-                        },
-                        ComponentStatus::Escaped => self.editor = None,
-                    }
-                    Err(_) => {},
-                }
-                EditorEventResult::KeepActive
-            } else if self.highlight {
-                match key_event.code {
-                    KeyCode::Esc => EditorEventResult::Escaped,
-                    KeyCode::Down => {
-                        self.highlight = false;
-                        EditorEventResult::ExitBottom
-                    }
-                    KeyCode::Up => {
-                        self.highlight = false;
-                        EditorEventResult::ExitTop
-                    }
-                    KeyCode::Enter => {
-                        let mut editor = TextInput::new(42, false);
-                        editor.set_text(self.node_id.clone());
-                        self.editor = Some(editor);
-                        EditorEventResult::KeepActive
-                    }
-                    _ => EditorEventResult::KeepActive,
-                }
-            } else {
-                EditorEventResult::Inactive
-            }
-        }
-
-        fn calculate_render_height(&self) -> usize {
-            1
-        }
-
-        fn get_text_output(&self, text: &mut String) {
-            writeln!(text, "NodeId: {}", self.node_id).unwrap();
-        }
-
-        fn get_highlight_prefix(&self) -> Option<usize> {
-            if self.highlight {
-                Some(8)
-            } else {
-                None
-            }
-        }
-
-        fn get_editor(&mut self) -> Option<&mut TextInput> {
-            self.editor.as_mut()
-        }
-
-        fn get_parse_error(&mut self) -> Option<&mut ModalMessage> {
-            self.parse_error.as_mut()
-        }
-    }
-}
-
-pub mod permission {
-    use super::*;
-
-    use std::collections::HashSet;
-
-    use golem_certificate::schemas::permissions::{Permissions, PermissionDetails, OutboundPermissions};
-    use url::Url;
-
-    pub struct PermissionEditor {
-        highlight: Option<usize>,
-        permissions: Permissions,
-        urls: Vec<Url>,
-        url_editor: Option<TextInput>,
-        parse_error: Option<ModalMessage>,
-    }
-
-    impl PermissionEditor {
-        pub fn new(permissions: Option<Permissions>) -> Self {
-            let default_permissions =
-                Permissions::Object(PermissionDetails {
-                    outbound: Some(OutboundPermissions::Urls(HashSet::new()))
-                });
-            let mut urls = vec![];
-            Self {
-                highlight: None,
-                permissions: permissions.map(|p| match &p {
-                    Permissions::Object(PermissionDetails { outbound: None }) => default_permissions.clone(),
-                    Permissions::Object(PermissionDetails { outbound: Some(OutboundPermissions::Urls(url_set)) }) => {
-                        url_set.iter().for_each(|url| urls.push(url.to_owned()));
-                        urls.sort();
-                        p
-                    }
-                    _ => p,
-                }).unwrap_or(default_permissions),
-                urls,
-                url_editor: None,
-                parse_error: None,
-            }
-        }
-
-        fn get_permissions(&self) -> Permissions {
-            match &self.permissions {
-                Permissions::Object(PermissionDetails { outbound: Some(OutboundPermissions::Urls(_)) }) => {
-                    let urls = self.urls.iter().map(|url| url.to_owned()).collect();
-                    Permissions::Object(PermissionDetails { outbound: Some(OutboundPermissions::Urls(urls)) })
-                }
-                p => p.clone(),
-            }
-        }
-    }
-
-    impl Default for PermissionEditor {
-        fn default() -> Self {
-            Self::new(None)
-        }
-    }
-
-    impl EditorComponent for PermissionEditor {
-        fn enter_from_below(&mut self) {
-            self.highlight = Some(self.calculate_render_height() - 1);
-        }
-
-        fn enter_from_top(&mut self) {
-            self.highlight = Some(0);
-        }
-
-        fn get_highlight(&self) -> Option<usize> {
-            self.highlight.clone()
-        }
-
-        fn handle_key_event(&mut self, key_event: KeyEvent) -> EditorEventResult {
-            let render_height = self.calculate_render_height();
-            if let Some(parse_error) = self.parse_error.as_mut() {
-                match parse_error.handle_key_event(key_event) {
-                    Ok(status) => match status {
-                        ComponentStatus::Active => {},
-                        _ => self.parse_error = None,
-                    }
-                    Err(_) => {},
-                }
-                EditorEventResult::KeepActive
-            } else if let Some(editor) = self.url_editor.as_mut() {
-                match editor.handle_key_event(key_event) {
-                    Ok(status) => match status {
-                        ComponentStatus::Active => {},
-                        ComponentStatus::Closed => {
-                            match Url::parse(editor.get_text()) {
-                                Ok(url) => {
-                                    let idx = self.highlight.as_ref().unwrap() - 3;
-                                    if idx == self.urls.len() {
-                                        self.urls.push(url);
-                                    } else {
-                                        self.urls[idx] = url;
-                                    }
-                                    self.url_editor = None;
-                                },
-                                Err(err) => self.parse_error = Some(ModalMessage::new("Url parse error", err.to_string())),
-                            }
-                        },
-                        ComponentStatus::Escaped => self.url_editor = None,
-                    }
-                    Err(_) => {},
-                }
-                EditorEventResult::KeepActive
-            } else if let Some(highlight) = self.highlight {
-                match key_event.code {
-                    KeyCode::Esc => EditorEventResult::Escaped,
-                    KeyCode::Down => {
-                        if highlight < render_height - 1 {
-                            let inc = if highlight == 1 { 2 } else { 1 };
-                            self.highlight = Some(highlight + inc);
-                            EditorEventResult::KeepActive
-                        } else {
-                            self.highlight = None;
-                            EditorEventResult::ExitBottom
-                        }
-                    }
-                    KeyCode::Up => {
-                        if highlight > 0 {
-                            let dec = if highlight == 3 { 2 } else { 1 };
-                            self.highlight = Some(highlight - dec);
-                            EditorEventResult::KeepActive
-                        } else {
-                            self.highlight = None;
-                            EditorEventResult::ExitTop
-                        }
-                    }
-                    KeyCode::Enter => {
-                        if highlight == 0 {
-                            self.permissions = match &self.permissions {
-                                Permissions::All => Permissions::Object(PermissionDetails { outbound: Some(OutboundPermissions::Unrestricted) }),
-                                Permissions::Object(_) => Permissions::All,
-                            };
-                        } else if highlight == 1 {
-                            match &mut self.permissions {
-                                Permissions::All => unreachable!("permission editor internal state error"),
-                                Permissions::Object(PermissionDetails { outbound }) =>
-                                    match outbound.as_ref().unwrap() {
-                                        OutboundPermissions::Unrestricted =>
-                                            outbound.insert(OutboundPermissions::Urls(Default::default())),
-                                        OutboundPermissions::Urls(_) =>
-                                            outbound.insert(OutboundPermissions::Unrestricted),
-                                    },
-                            };
-                        } else if highlight > 2 && highlight <= render_height {
-                            let idx = highlight - 3;
-                            let mut editor = TextInput::new(255, false);
-                            if idx < self.urls.len() {
-                                editor.set_text(self.urls[idx].to_string());
-                            }
-                            self.url_editor = Some(editor);
-                        }
-                        EditorEventResult::KeepActive
-                    }
-                    KeyCode::Delete | KeyCode::Backspace => {
-                        if highlight > 2 && highlight < render_height {
-                            let idx = highlight - 3;
-                            self.urls.remove(idx);
-                            if highlight > 3 {
-                                self.highlight = Some(highlight - 1);
-                            }
-                        }
-                        EditorEventResult::KeepActive
-                    }
-                    _ => EditorEventResult::KeepActive,
-                }
-            } else {
-                EditorEventResult::Inactive
-            }
-        }
-
-        fn calculate_render_height(&self) -> usize {
-            match &self.permissions {
-                Permissions::All => 1,
-                Permissions::Object(PermissionDetails { outbound }) =>
-                    match outbound.as_ref() {
-                        Some(outbound_details) => match outbound_details {
-                            OutboundPermissions::Unrestricted => 2,
-                            OutboundPermissions::Urls(_) => 4 + self.urls.len(),
-                        }
-                        None => 1,
-                    }
-            }
-        }
-
-        fn get_text_output(&self, text: &mut String) {
-            write!(text, "Permissions").unwrap();
-            match &self.permissions {
-                Permissions::All => writeln!(text, ": All").unwrap(),
-                Permissions::Object(PermissionDetails { outbound: None }) => writeln!(text, "None").unwrap(),
-                Permissions::Object(PermissionDetails { outbound: Some(outbound_details) }) => {
-                    writeln!(text, "").unwrap();
-                    write!(text, "  Outbound").unwrap();
-                    match outbound_details {
-                        OutboundPermissions::Unrestricted => {
-                            writeln!(text, ": Unrestricted").unwrap();
-                        }
-                        OutboundPermissions::Urls(_) => {
-                            writeln!(text, "").unwrap();
-                            writeln!(text, "    Urls").unwrap();
-                            self.urls.iter()
-                                .for_each(|url| writeln!(text, "      {}", url).unwrap());
-                        }
-                    }
-                }
-            }
-        }
-
-        fn get_highlight_prefix(&self) -> Option<usize> {
-            self.highlight.map(|highlight| {
-                match highlight {
-                    0 => 0,
-                    1 => 2,
-                    _ => 6,
-                }
-            })
-        }
-
-        fn get_editor(&mut self) -> Option<&mut TextInput> {
-            self.url_editor.as_mut()
-        }
-
-        fn get_parse_error(&mut self) -> Option<&mut ModalMessage> {
-            self.parse_error.as_mut()
-        }
-    }
-}
-
-pub mod validity_period {
-    use super::*;
-
-    use chrono::{DateTime, Utc};
-    use golem_certificate::schemas::validity_period::ValidityPeriod;
-
-    pub struct ValidityPeriodEditor {
-        not_before: DateTime<Utc>,
-        not_after: DateTime<Utc>,
-        highlight: Option<usize>,
-        date_editor: Option<TextInput>,
-        parse_error: Option<ModalMessage>,
-    }
-
-    impl ValidityPeriodEditor {
-        pub fn new(validity_period: Option<ValidityPeriod>) -> Self {
-            let (not_before, not_after) = match validity_period {
-                Some(ValidityPeriod { not_before, not_after }) => (not_before, not_after),
-                None => (Utc::now(), Utc::now()),
-            };
-            Self {
-                not_before,
-                not_after,
-                highlight: None,
-                date_editor: None,
-                parse_error: None,
-            }
-        }
-
-        pub fn get_validity_period(&self) -> ValidityPeriod {
-            ValidityPeriod {
-                not_before: self.not_before.clone(),
-                not_after: self.not_after.clone(),
-            }
-        }
-    }
-
-    impl Default for ValidityPeriodEditor {
-        fn default() -> Self {
-            Self::new(None)
-        }
-    }
-
-    impl EditorComponent for ValidityPeriodEditor {
-        fn enter_from_below(&mut self) {
-            self.highlight = Some(2);
-        }
-
-        fn enter_from_top(&mut self) {
-            self.highlight = Some(1);
-        }
-
-        fn get_highlight(&self) -> Option<usize> {
-            self.highlight
-        }
-
-        fn handle_key_event(&mut self, key_event: KeyEvent) -> EditorEventResult {
-            if let Some(parse_error) = self.parse_error.as_mut() {
-                match parse_error.handle_key_event(key_event) {
-                    Ok(status) => match status {
-                        ComponentStatus::Active => {},
-                        _ => self.parse_error = None,
-                    }
-                    Err(_) => {},
-                };
-                EditorEventResult::KeepActive
-            } else if let Some(date_editor) = self.date_editor.as_mut() {
-                match date_editor.handle_key_event(key_event) {
-                    Ok(status) => match status {
-                        ComponentStatus::Active => {},
-                        ComponentStatus::Closed => {
-                            match date_editor.get_text().parse::<DateTime<Utc>>() {
-                                Ok(datetime) => {
-                                    if self.highlight.unwrap() == 1 {
-                                        self.not_before = datetime;
-                                    } else {
-                                        self.not_after = datetime;
-                                    }
-                                    self.date_editor = None;
-                                }
-                                Err(err) => {
-                                    let error = ModalMessage::new("Datetime parse error", err.to_string());
-                                    self.parse_error = Some(error);
-                                }
-                            }
-
-                        }
-                        ComponentStatus::Escaped => self.date_editor = None,
-                    }
-                    Err(_) => {},
-                };
-                EditorEventResult::KeepActive
-            } else if let Some(highlight) = self.highlight {
-                match key_event.code {
-                    KeyCode::Esc => EditorEventResult::Escaped,
-                    KeyCode::Down =>
-                        if highlight == 2 {
-                            self.highlight = None;
-                            EditorEventResult::ExitBottom
-                        } else {
-                            self.highlight = Some(2);
-                            EditorEventResult::KeepActive
-                        },
-                    KeyCode::Up =>
-                        if highlight == 1 {
-                            self.highlight = None;
-                            EditorEventResult::ExitTop
-                        } else {
-                            self.highlight = Some(1);
-                            EditorEventResult::KeepActive
-                        },
-                    KeyCode::Enter => {
-                        // 2014-11-28T21:00:09+09:00 => 25
-                        let mut editor = TextInput::new(30, false);
-                        if highlight == 1 {
-                            editor.set_text(self.not_before.to_string());
-                        } else {
-                            editor.set_text(self.not_after.to_string());
-                        }
-                        self.date_editor = Some(editor);
-                        EditorEventResult::KeepActive
-                    },
-                    _ => EditorEventResult::KeepActive,
-                }
-            } else {
-                EditorEventResult::Inactive
-            }
-        }
-
-        fn calculate_render_height(&self) -> usize  {
-            3
-        }
-
-        fn get_text_output(&self, text: &mut String) {
-            writeln!(text, "Validity Period").unwrap();
-            writeln!(text, "  Not Before: {}", self.not_before).unwrap();
-            writeln!(text, "  Not After:  {}", self.not_after).unwrap();
-        }
-
-        fn get_highlight_prefix(&self) -> Option<usize> {
-            self.highlight.map(|_| 14)
-        }
-
-        fn get_editor(&mut self) -> Option<&mut TextInput> {
-            self.date_editor.as_mut()
-        }
-
-        fn get_parse_error(&mut self) -> Option<&mut ModalMessage> {
-            self.parse_error.as_mut()
-        }
     }
 }
